+++ ps.sh
@@ -2,9 +2,9 @@
 ##
 # check what is going on ;)
 ##
-if ps -ef | grep -q "[s]crape\|[p]arse\|[m]arkdown"; then
-  process_name=$(ps -ef | grep -o '[s]crape\|[p]arse\|[m]arkdown')
-  elapsed_seconds=$(ps -p $(pgrep -f "python3 ransomwatch.py $process_name") -o etimes= | awk '{print $1}')
+if ps -ef | grep ransomwatch | grep -q "[s]crape\|[p]arse\|[m]arkdown"; then
+  process_name=$(ps -ef | grep ransomwatch | grep -o '[s]crape\|[p]arse\|[m]arkdown')
+  elapsed_seconds=$(ps -p $(pgrep -f "python3 ransomwatchv2.py $process_name") -o etimes= | awk '{print $1}')
   elapsed_minutes=$((elapsed_seconds/60))
   bold=$(tput bold)
   reset=$(tput sgr0)