+++ sharedutils.py
@@ -335,7 +335,6 @@ def grouppostavailable(groupname):
     else:
         return False
 
-
 def grouppostcount(groupname):
     grouppost_count = 0
     posts = openjson('posts.json')
@@ -348,6 +347,14 @@ def grouppostcount(groupname):
         grouppost_count = '1 post found'
     elif grouppost_count ==0:
         grouppost_count = 'no post found'
+    return grouppost_count
+
+def postcountgroup(groupname):
+    grouppost_count = 0
+    posts = openjson('posts.json')
+    for post in posts:
+        if post['group_name'] == groupname:
+            grouppost_count += 1
     return grouppost_count
 
 def groupcount():